requirements.txt
python-dotenv==1.0.0
groq
AppOpener
pywhatkit
bs4
pillow
rich==13.7.0
requests
keyboard
cohere==4.39
googlesearch-python==1.2.3
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
gunicorn==21.2.0
flask
urllib3>=2.0.7